viz: added working play state buttons
Added play/step/pause/stop buttons to html, along
with a display of the current stepnum. These were wired
into a new set of exported javascript functions which
update an atop with the current playstate.

Internal to the simulation, this atom is monitored
and the appropiate action is taken. This included
splitting out the initialization and tracking
the current state of the sim.

Also added a separate binding for the step display
(currently a debug print of the sim-state).